# `http_solver`: optimal graph algo using BFS