--- a/crates/autopilot/src/arguments.rs
+++ b/crates/autopilot/src/arguments.rs
@@ -349,6 +349,12 @@ pub struct FeePolicy {
     /// - Surplus with cap:
     /// surplus:0.5:0.06
     ///
+    /// - Price improvement without cap:
+    /// price_improvement:0.5:1.0
+    ///
+    /// - Price improvement with cap:
+    /// price_improvement:0.5:0.06
+    ///
     /// - Volume based:
     /// volume:0.1
     #[clap(long, env, default_value = "surplus:0.0:1.0")]
@@ -371,6 +377,7 @@ impl FeePolicy {
                 factor,
                 max_volume_factor,
             },
+            FeePolicyKind::PriceImprovement { .. } => todo!(),
             FeePolicyKind::Volume { factor } => domain::fee::Policy::Volume { factor },
         }
     }
@@ -380,6 +387,10 @@ impl FeePolicy {
 pub enum FeePolicyKind {
     /// How much of the order's surplus should be taken as a protocol fee.
     Surplus { factor: f64, max_volume_factor: f64 },
+    /// How much of the order's price improvement should be taken as a protocol
+    /// fee where price improvement is a difference between the executed price
+    /// and the best quote.
+    PriceImprovement { factor: f64, max_volume_factor: f64 },
     /// How much of the order's volume should be taken as a protocol fee.
     Volume { factor: f64 },
 }
@@ -403,6 +414,22 @@ impl FromStr for FeePolicyKind {
                     .parse::<f64>()
                     .map_err(|e| format!("invalid max volume factor: {}", e))?;
                 Ok(Self::Surplus {
+                    factor,
+                    max_volume_factor,
+                })
+            }
+            "priceImprovement" => {
+                let factor = parts
+                    .next()
+                    .ok_or("missing price improvement factor")?
+                    .parse::<f64>()
+                    .map_err(|e| format!("invalid price improvement factor: {}", e))?;
+                let max_volume_factor = parts
+                    .next()
+                    .ok_or("missing price improvement max volume factor")?
+                    .parse::<f64>()
+                    .map_err(|e| format!("invalid price improvement max volume factor: {}", e))?;
+                Ok(Self::PriceImprovement {
                     factor,
                     max_volume_factor,
                 })