### Description
<!--- Describe your changes to provide context for reviewers, including why it is needed -->
Fixes a todo: earlier, the code used a more iterative approach with HashSet manipulations
Now: we construct a graph and use BFS to explore connected tokens.
Hopefully this is more optimal graph algorithm.

### Changes
<!-- List of detailed changes (how the change is accomplished) -->

Implemented a graph data structure and BFS to efficiently find tokens connected to the native token.
## How to test
Tests passed (run using):
```sh
cargo test --package solver --lib -- solver::http_solver::instance_creation::tests --nocapture
```                                                                      

<!--
## Related Issues

Fixes #
-->