@wthrajat 2024-02-13
@MartinquaXD Not sure of the best way to benchmark it. I used [hyperfine](https://github.com/sharkdp/hyperfine) for benchmarking unit tests times for this branch and `main` branch.
And got the following results:

#### `main` branch:
```sh
~/rust/cowprotocol/services  on main
╰─❯ hyperfine 'cargo test --package solver --lib -- solver::http_solver::instance_creation::tests --nocapture' ─╯
Benchmark 1: cargo test --package solver --lib -- solver::http_solver::instance_creation::tests --nocapture

  Time (mean ± σ):     11.784 s ± 36.549 s    [User: 55.875 s, System: 2.913 s]
  Range (min … max):    0.220 s … 115.804 s    10 runs
```
Screenshot:
![2024-02-13-01 40 45-screenshot](https://github.com/cowprotocol/services/assets/38693805/6ffd9e41-eb90-403d-aa54-a92ef625b0be)
<hr>


#### `optimal_graph_algo` (this) branch:

```sh
~/rust/cowprotocol/services  on optimal_graph_algo
╰─❯ hyperfine 'cargo test --package solver --lib -- solver::http_solver::instance_creation::tests --nocapture'                      ─╯
Benchmark 1: cargo test --package solver --lib -- solver::http_solver::instance_creation::tests --nocapture

  Time (mean ± σ):     10.397 s ± 32.152 s    [User: 39.119 s, System: 2.047 s]
  Range (min … max):    0.222 s … 101.904 s    10 runs
```
Screenshot:
![2024-02-13-15 21 57-screenshot](https://github.com/cowprotocol/services/assets/38693805/f62caad0-5258-4d6c-bd71-1a5abbb7f493)

<hr>
Both the tests were run after a fresh build, eliminating the chance of using cache.

### Result

There is a slight time latency improvement after the PR.
1. The maximum time has been improved by `13.900 sec`
```math
115.804 - 101.904 =  13.900
3. If considering `mean time`, this has been improved by `11.77%`
```math
((11.784 − 10.397​) / 11.784) * 100
```
However, I'm not too sure how to test the memory usage concretely. I could take a look at the memory usage during both the test runs but that would