@wthrajat 2024-02-13
@MartinquaXD Not sure of the best way to benchmark it. I used [hyperfine](https://github.com/sharkdp/hyperfine) for benchmarking unit tests times for this branch and main branch.
And got the following results on `12 runs` sample:

#### `main` branch:
![2024-02-13-16 51 39-screenshot](https://github.com/cowprotocol/services/assets/38693805/b394e07c-98a1-410a-8099-7466789f19e7)

<hr>

#### `this` branch:
![2024-02-13-16 52 34-screenshot](https://github.com/cowprotocol/services/assets/38693805/b1238b1a-19dd-4d16-83e3-d0fafc4eeff3)

<hr>

### Result

There are some noticeable time improvements:

1. `minimum time` has gone down from `226 ms` to `215 ms`
2. `maximum time` has gone down from `264 ms` to `236 ms`
3. `mean time` extremes have been improved by *5.1%*
```math
\text{} \frac{237.6 - 225.5}{237.6} \times 100 = 5.1 %
```

However, I'm not too sure how to test the memory usage concretely (AFAIK, [criterion](https://docs.rs/criterion/latest/criterion) doesn't support memory usage). I could take a look at the memory usage during both the test runs (on a quick look, doesn't seem to be a noticeable difference) but that would not be too accurate. If you have any ideas, feel free to suggest, I'll give them a shot.

Also if the original TODO meant some other optimization, feel free to point me to that.
Thanks!