@MartinquaXD 2024-02-13
> Not sure of the best way to benchmark it.

This is not so easy. I just tried to do it but it turns out this code is actually not used anymore. The `solver` crate is currently being phased out. We have not been running it for some time but can't delete it yet because some other crates are reusing code from it. I was under the impression the code was still in use by a different component but that is not the case.
Sorry that this ended up wasting your time. :/
Will open a PR to get rid of the unused code so nobody else runs into the same issue.