@wthrajat 2024-02-14
@MartinquaXD 
> Sorry that this ended up wasting your time. :/

No it's chill, I'm in the learning phase so it doesn't hurt.
I see that `solver` crate has been removed by the PR #2405. So I guess I should close this now.
Thanks for the review!